:recycle: move logging package to router/
